+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 2. Added Signal Messanger
 3. Fixed waybar
 4. Removed keyboard state from waybar
+5. Added wlogout config